Add `insecure` option
This option, defaulted to false, passes `-k` to curl and allows it to
work in cases where the website being checked is deployed with
self-signed certificates in CI.

`insecure` is inserted as the first argument to the script to minimize
unexpected behavior in case the scan_text has whitespace that has not
been escaped by mistake. Consider the following example:

scan-website.sh https://example.com/ text with whitespace true

Had we used insecure=$3, we'd have ended up with insecure="with",
which would run curl without `-k` and presumably fail in a mysterious
way.